refactor useScrollAnchor hook
